# perf: cache parsed `assets_json` instead of string