So that it needn't be parsed every time.

## Performance of cached load

### Before

```python
In [3]: %timeit frappe.utils.get_assets_json()
24 µs ± 154 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```

### After (96% better)

```python
In [3]: %timeit frappe.utils.get_assets_json()
841 ns ± 7.35 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
```