@ankush 2022-04-01
```
line 33, in block 'head'
15:50:31 web.1            |     {{ include_style('website.bundle.css') }}
15:50:31 web.1            |   File "/Users/hussain/Frappe/cloud-bench/env/lib/python3.8/site-packages/jinja2/sandbox.py", line 393, in call
15:50:31 web.1            |     return __context.call(__obj, *args, **kwargs)
15:50:31 web.1            |   File "/Users/hussain/Frappe/cloud-bench/apps/frappe/frappe/utils/jinja_globals.py", line 84, in include_style
15:50:31 web.1            |     path = bundled_asset(path)
15:50:31 web.1            |   File "/Users/hussain/Frappe/cloud-bench/apps/frappe/frappe/utils/jinja_globals.py", line 93, in bundled_asset
15:50:31 web.1            |     bundled_assets = get_assets_json()
15:50:31 web.1            |   File "/Users/hussain/Frappe/cloud-bench/apps/frappe/frappe/utils/__init__.py", line 805, in get_assets_json
15:50:31 web.1            |     frappe.local.assets_json = frappe.cache().get_value(
15:50:31 web.1            |   File "/Users/hussain/Frappe/cloud-bench/apps/frappe/frappe/utils/redis_wrapper.py", line 76, in get_value
15:50:31 web.1            |     val = pickle.loads(val)
15:50:31 web.1            | _pickle.UnpicklingError: invalid load key, '{'.
```

probably related? 🤔