@sagarvora 2022-04-01
Also might be worth checking if `frappe.cache().flushall()` solves the issue.